requirements.txt
flask==2.3.2
gunicorn==20.1.0
tweepy==4.12.1
requests==2.28.2
pandas==1.5.3
beautifulsoup4==4.11.2
selenium==4.8.2
schedule==1.1.0
google-cloud-storage==2.8.0
google-cloud-firestore==2.11.0
firebase-admin==6.1.0
python-dotenv==1.0.0
flask-login==0.6.2
werkzeug==2.3.6
flask-wtf==1.1.1